+++ radio/src/gui/480x272/model_setup.cpp
@@ -118,7 +118,7 @@ void editTimerMode(int timerIdx, coord_t y, LcdFlags attr, event_t event)
     switch (menuHorizontalPosition) {
       case 0:
       {
-        int8_t timerMode = timer.mode;
+        int32_t timerMode = timer.mode;
         if (timerMode < 0) timerMode -= TMRMODE_COUNT-1;
         CHECK_INCDEC_MODELVAR_CHECK(event, timerMode, -TMRMODE_COUNT-SWSRC_LAST+1, TMRMODE_COUNT+SWSRC_LAST-1, isSwitchAvailableInTimers);
         if (timerMode < 0) timerMode += TMRMODE_COUNT-1;
